apps/raspi_os: Remove mbedtls config flag from sdkconfig
With the updation of mbedtls library to version 2.16.0, the config
option 'MBEDTLS_SSL_MAX_CONTENT_LEN' is no longer used in mbedtls
source code. Hence user does not require the config option in sdkconfig.

Also, 'MBEDTLS_SSL_IN_CONTENT_LEN' and 'MBEDTLS_SSL_OUT_CONTENT_LEN'
flags are used in this version, and are already defined in config.h
with values of '6144' and '4096' respectively.